Function that prints 10 times the numbers, from 0 to 14

+++ 0x04-more_functions_nested_loops/5-more_numbers.c
@@ -4,9 +4,9 @@
  **/
 void more_numbers(void)
 {
-	int i, j;
-	int range = 14;
-	int times = 10;
+	short i, j;
+	short range = 14;
+	short times = 10;
 
 	for (i = 0; i < times; i++)
 	{
@@ -20,6 +20,6 @@ void more_numbers(void)
 			else
 				_putchar('0' + j);
 		}
-	_putchar('\n');
+		_putchar('\n');
 	}
 }
